requirements.txt
numpy>=1.22
cupy-cuda12x>=12.0
scikit-learn>=1.0
h5py>=3.8
